backend/requirements.txt
# --- Framework principal ---
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.11.0
pydantic==2.12.3
email-validator==2.3.0

# --- Base de datos MongoDB (Async) ---
pymongo[srv]==4.10.1
motor==3.7.1

# --- Seguridad y autenticación ---
passlib==1.7.4
pyjwt==2.10.1
bcrypt==4.2.0
python-dotenv==1.0.1


# --- IA / Embeddings / NLP ---
sentence-transformers==2.7.0
torch==2.9.0
numpy==2.3
scikit-learn==1.6.0 

# API OpenAI
openai==1.97.1

# --- Procesamiento de texto / OCR / PDF ---
pymupdf==1.26.5
pdf2image==1.17.0
pytesseract==0.3.13
pillow==10.4.0

# --- Utilidades / herramientas ---
requests==2.32.3
aiofiles==24.1.0
loguru==0.7.2

# --- Email / notificaciones (si usas envío) ---
aiosmtplib==3.0.2

# --- Testeo y dev ---
pytest==8.3.3
httpx==0.27.2
